refactor: 로그인 로직 분리로 인한 테스트 로직 분리 1. 로그인 관련 테스트들은 LoginServiceIntegrationTest로 분리 2. 회원 ID 검증 값이 auto_increament로 인한 Assertion 에러 발생 -> setUP 메소드로 id가 1부터 시작하도록 변경
