+++ src/test/java/com/sns/whisper/spring/integration/user/UserServiceIntegrationTest.java
@@ -4,21 +4,19 @@ package com.sns.whisper.spring.integration.user;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import com.sns.whisper.common.factory.UserFactory;
 import com.sns.whisper.domain.user.application.UserService;
 import com.sns.whisper.domain.user.application.dto.request.UserSignUpServiceRequest;
 import com.sns.whisper.domain.user.application.dto.response.UserResponse;
-import com.sns.whisper.domain.user.application.session.SessionManager;
 import com.sns.whisper.domain.user.domain.User;
 import com.sns.whisper.domain.user.domain.profile.UserStatus;
 import com.sns.whisper.domain.user.infrastructure.JPAUserRepository;
 import com.sns.whisper.exception.user.DuplicatedUserIdException;
-import com.sns.whisper.exception.user.LoginFailException;
-import com.sns.whisper.global.common.PasswordEncryptor;
 import com.sns.whisper.spring.integration.IntegrationTest;
+import jakarta.persistence.EntityManager;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,7 +32,15 @@ public class UserServiceIntegrationTest extends IntegrationTest {
     private JPAUserRepository userRepository;
 
     @Autowired
-    private SessionManager sessionManager;
+    private EntityManager entityManager;
+
+    @BeforeEach
+    void setUp() {
+        entityManager
+                .createNativeQuery(
+                        "ALTER TABLE \"user\" ALTER COLUMN `id` RESTART WITH 1")
+                .executeUpdate();
+    }
 
     @AfterEach
     void tearDown() {
@@ -93,63 +99,6 @@ public class UserServiceIntegrationTest extends IntegrationTest {
                                                          .hasFieldOrPropertyWithValue("httpStatus",
                                                                  HttpStatus.BAD_REQUEST)
                                                          .hasMessage("중복된 아이디입니다.");
-    }
-
-    @Test
-    @DisplayName("가입 회원의 유효한 회원 아이디와 비밀번호를 입력하면, 로그인할 수 있다.")
-    void login_ValidUser_Success() throws Exception {
-        //given
-        String userId = "userId12";
-        String password = "password1234";
-
-        User user = UserFactory.createBasicUser(userId, password);
-        userRepository.save(user);
-
-        //when
-        userService.login(userId, password);
-
-        //then
-        assertThat(sessionManager.extractUser("ACCESS_USER")).isEqualTo(userId);
-        assertThat(PasswordEncryptor.isMatch(password, user.getPassword())).isTrue();
-
-    }
-
-    @Test
-    @DisplayName("가입하지 않은 회원은 로그인을 할 수 없다.")
-    void login_NotValidUserId_ExceptionThrown() throws Exception {
-        //given
-        String userId = "userId123";
-        String password = "password1234";
-
-        //when, then
-        assertThatCode(() -> userService.login(userId, password))
-                .isInstanceOf(
-                        LoginFailException.class)
-                .hasFieldOrPropertyWithValue(
-                        "httpStatus",
-                        HttpStatus.BAD_REQUEST)
-                .hasMessage("아이디와 비밀번호를 확인해주세요.");
-
-    }
-
-    @Test
-    @DisplayName("유효하지 않은 비밀번호를 입력할 경우 로그인을 할 수 없다.")
-    void login_NotValidPassword_ExceptionThrown() throws Exception {
-        //given
-        String userId = "userId123";
-        String wrongPassword = "wrongPass1234";
-
-        User user = UserFactory.createBasicUser(userId, "password1234");
-        userRepository.save(user);
-
-        //when, then
-        assertThatCode(() -> userService.login(userId, wrongPassword))
-                .isInstanceOf(
-                        LoginFailException.class)
-                .hasFieldOrPropertyWithValue(
-                        "httpStatus",
-                        HttpStatus.BAD_REQUEST)
-                .hasMessage("아이디와 비밀번호를 확인해주세요.");
     }
 
 }
